+++ tests/cache/test_cache.py
@@ -21,7 +21,6 @@ class ClownIntel(BaseIntel):
     juggles: bool
     wig: WigIntel
 
-
 test_limit = 100
 
 
@@ -91,4 +90,27 @@ class TestCache(TestCase):
 
     def test_sqlite_cache_limit(self):
         self.run_text_cache_limit(sql_lite_cache, cache_to_sqlite)
+
+    def run_complex_object_cache(self, cache: BaseCache, cache_decorator: Callable):
+        cache.clear()
+
+        class Candy:
+            def __init__(self, sweetness: int):
+                self.sweetness = sweetness
+
+        @cache_decorator(limit=test_limit * 3)
+        def create_candy(candy: Candy) -> Candy:
+            return Candy(sweetness=candy.sweetness)
+
+        for _ in range(3):
+            for i in range(test_limit):
+                _ = create_candy(Candy(sweetness=i))
+
+        self.assertEqual(len(cache), test_limit)
+
+    def test_memory_complex_object_cache(self):
+        self.run_complex_object_cache(memory_cache, cache_to_memory)
+
+    def test_sqlite_complex_object_cache(self):
+        self.run_complex_object_cache(sql_lite_cache, cache_to_sqlite)